requirements.txt
fastapi==0.115.6
langchain==0.3.14
langchain-openai==0.3.1
langchain-pinecone==0.2.2
openai==1.59.8
pinecone==5.4.2
# pinecone-client==5.0.1
transformers==4.48.0
torch==2.5.1
tiktoken==0.8.0
uvicorn==0.34.0
pydantic==2.10.5
python-dotenv==1.0.1  # If you use environment variables
python-telegram-bot==21.10
